pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omarchy-mcp"
version = "0.0.1"
description = "MCP Server for Omarchy"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "ServiceStack", email = "team@servicestack.net"}
]
keywords = ["mcp", "omarchy", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastmcp>=0.1.0",
    "google-genai>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/ServiceStack/omarchy-mcp"
Repository = "https://github.com/ServiceStack/omarchy-mcp"

[project.scripts]
omarchy-mcp = "omarchy_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/omarchy_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
